+++ frameworks/jackson-bench/src/main/java/com/github/skjolber/nve/jackson/ReducedJacksonParser.java
@@ -14,9 +14,9 @@ import com.fasterxml.jackson.module.afterburner.AfterburnerModule;
 
 public class ReducedJacksonParser {
 
-    public static int countDefCveItem(List<byte[]> contents) throws IOException {
+    public static long countDefCveItem(List<byte[]> contents) throws IOException {
 
-		int count = 0;
+    	long count = 0;
 
 		ObjectMapper objectMapper = new ObjectMapper().configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
 		objectMapper.registerModule(new AfterburnerModule());
